app/requirements.txt
streamlit
opencv-python
python-dotenv
google-generativeai
pytesseract
Pillow
numpy
sqlite3
os
pdf2image
poppler-utils
zstd
ollama
imagehash
Pillow
spacy
